[logs>debug] Merge logs into debug
